refactor: extract lifecycle builder from api/app.py
